% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction_triad.R
\name{bias_correction_triad}
\alias{bias_correction_triad}
\title{Bias Correction Triad}
\usage{
bias_correction_triad(songs, regions, group_year = 1L, verbose = TRUE)
}
\arguments{
\item{songs}{a tibble of songs.}
}
\value{
\code{bias_correction_triad} returns a tibble, with time and space weights for each song, taking into account triad weight.
}
\description{
Bias Correction Triad
}
\examples{
w_time_space2 <- bias_correction_triad(songs, regions, group_year = 1L)
}
