% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_triad_from_config.R
\name{build_triad_from_config}
\alias{build_triad_from_config}
\title{Get Triad Type Table}
\usage{
build_triad_from_config(env = "spotify")
}
\arguments{
\item{env}{a character string}
}
\value{
\code{build_triad_from_config} returns a triad table.
}
\description{
Read config file and list all triads#'
}
\examples{
tabTriads <- build_triad_from_config(env = "spotify")
}
