% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_song_title.R
\name{clean_song_title}
\alias{clean_song_title}
\title{Clean Song Names}
\usage{
clean_song_title(.data)
}
\arguments{
\item{.data}{a data frame with two columns Song and Artist}
}
\value{
\code{clean_song_title} returns a tibble with a modified column Song
}
\description{
Remove "Remaster version" and "featuring..." mentions.
}
\examples{
tibble::tibble(Song = c("You're So Sweet Darling, BWV 565: I. Toccata",
                        "That's All Right - Remaster 2006 definitive edition"),
               Artist = c("Belzebuth & The Jeremies", "Benny Goodman")) \%>\% 
  clean_song_title()
}
