% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_population_estimates.R
\name{compute_population_estimates}
\alias{compute_population_estimates}
\title{Compute population Estimates}
\usage{
compute_population_estimates(update = TRUE, group_year = 1L)
}
\arguments{
\item{update}{a boolean}

\item{group_year}{an integer giving the number of year for grouping.}
}
\value{
\code{compute_population_estimates} returns a tibble of worldwide population estimates
}
\description{
Population estimates per year and per country.
}
\examples{
population <- compute_population_estimates(update = FALSE)
# population \%>\%
#   ggplot2::ggplot() +
#   ggplot2::aes(Year_Group, weight_year_country, group = country) +
#   ggplot2::geom_line()
}
