% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_id.R
\name{get_playlist_id}
\alias{get_playlist_id}
\title{Get Playlist id}
\usage{
get_playlist_id(triad = NULL, name = NULL)
}
\arguments{
\item{triad}{a character string (exemple : "1 7")}

\item{name}{a character string, it can be partial (exemple : "Carlton").}
}
\value{
\code{get_playlist_id} returns a Spotify playlist id
}
\description{
Get the spotify playlist id based on the name of the triad
}
\examples{
get_playlist_id(triad = "1 7")
get_playlist_id(name = "Vanessa Carlton")
}
