% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spotify_api_playlist.R
\name{get_spotify_api_playlist}
\alias{get_spotify_api_playlist}
\title{Get Tracks from Playlist Spotify}
\usage{
get_spotify_api_playlist(
  env = "spotify",
  playlist_id = NULL,
  market = NULL,
  fields = NULL,
  scope = c("public", "private"),
  verbose = TRUE
)
}
\arguments{
\item{env}{default to "spotify"}

\item{playlist_id}{a playlist id}

\item{market}{an additional parameter passed to the Spotify API}

\item{fields}{an additional parameter passed to the Spotify API}

\item{scope}{either "public" or "private".}

\item{verbose}{print message in console}
}
\value{
\code{get_spotify_api_playlist} returns a tibble with spotify data. 1 row per song and a featured artist.
}
\description{
Get Tracks from Playlist Spotify
}
\examples{
playlist_id <- get_playlist_id(triad = "1 7")
songs <- tvc:::get_spotify_api_playlist(
  playlist_id = playlist_id,
  scope = "public"
)
}
