% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_line_breaks_adaptive.R
\name{insert_line_breaks_adaptive}
\alias{insert_line_breaks_adaptive}
\title{Insert line breaks}
\usage{
insert_line_breaks_adaptive(texts, max_lines = 3L)
}
\arguments{
\item{texts}{a character vector}

\item{max_lines}{maximum number of lines expected for the longer character strings}
}
\value{
\code{insert_line_breaks_adaptive} returns \code{texts} with a proper amount of line breaks
}
\description{
Insert line breaks
}
\examples{
texts <- c("Bluebird",
           "Blue Jeans",
           "Blue Velvet",
           "Blue Moon (Remaster)",
           "Blue Banisters")
insert_line_breaks_adaptive(texts = texts, max_lines = 3L)
}
