% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_triade.R
\name{parse_triade}
\alias{parse_triade}
\title{parse_triade Title}
\usage{
parse_triade(data, songs)
}
\arguments{
\item{data}{a tibble of triads as read from googlesheet}

\item{songs}{a tibble of songs}
}
\value{
\code{parse_triade} returns a tibble of triads
}
\description{
parse_triade Title
}
\examples{
triads <- read_sheet_triad()
playlist_id <- get_playlist_id(triad = "1 7")
songs <- tvc:::get_spotify_api_playlist(playlist_id = playlist_id, scope = "public")
data <- parse_triade(triads, songs)
}
