% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_triad_coverage.R
\name{plot_triad_coverage}
\alias{plot_triad_coverage}
\title{Plot Triad Coverage}
\usage{
plot_triad_coverage(triads, weighted = TRUE)
}
\arguments{
\item{triads}{a tibble triad}

\item{weighted}{a boolean specifying if songs should be weighted using tvc's bias correction method}
}
\value{
\code{plot_triad_coverage} plots a triad coverage heatmap
}
\description{
Plot Triad Coverage
}
\examples{
triads <- read_sheet_triad()
plot_triad_coverage(triads = triads)
}
