% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tvc_histogram.R
\name{plot_tvc_histogram}
\alias{plot_tvc_histogram}
\title{Plot Occurences Histogram}
\usage{
plot_tvc_histogram(songs, triadType = "1 7", group_year = 1L, weighted = TRUE)
}
\arguments{
\item{songs}{a tibble with at least a column \code{Album_Year}, \code{Song} and \code{ISRC}.}

\item{triadType}{a character string specifying the triad.}
}
\value{
\code{plot_tvc_histogram} returns a plot
}
\description{
Plot Occurences Histogram
}
\examples{
songs <- tibble::tibble(Song = paste("Love Me", 1:30L, "Times"),
                        ISRC = sample(c(letters, LETTERS), size = 30L, replace = FALSE),
                        genres = purrr::map(1:30L,
                                            .f = ~ sample(x = c("rock", "metal",
                                                                "classical", "pop",
                                                                "yacht rock"),
                                                          size = 2)),
                        Album_Year = round(rnorm(n = 30L,
                                                 mean = 1986,
                                                 sd = 7)))
songs \%>\% 
  plot_tvc_histogram()
}
