% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tvc_network.R
\name{plot_tvc_network}
\alias{plot_tvc_network}
\title{Plot tvc songs network}
\usage{
plot_tvc_network(data)
}
\arguments{
\item{data}{a tibble as returned by \code{\link{parse_triade}}}
}
\value{
\code{plot_tvc_network} returns a network plot
}
\description{
Plot tvc songs network
}
\examples{
triads <- read_sheet_triad()
playlist_id <- get_playlist_id(triad = "1 7")
songs <- tvc:::get_spotify_api_playlist(playlist_id = playlist_id, scope = "public")
data <- parse_triade(triads, songs)
print(plot_tvc_network(data))
}
