% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheet_triad.R
\name{read_sheet_triad}
\alias{read_sheet_triad}
\title{Read Googlesheet Triad}
\usage{
read_sheet_triad()
}
\value{
\code{read_sheet_triad} returns a tibble with triad information
}
\description{
Read Googlesheet Triad
}
\examples{
triads <- read_sheet_triad()
}
