% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_tvc_c.R
\name{scale_colour_tvc_c}
\alias{scale_colour_tvc_c}
\title{Title}
\usage{
scale_colour_tvc_c(palette = "gradient", direction = 1, ...)
}
\arguments{
\item{palette}{either gradient or "main" or ?}

\item{direction}{1 for onward or -1 for downward}

\item{...}{other params ?}
}
\value{
\code{scale_colour_tvc_c} returns a corporate palette
}
\description{
Description
}
\examples{
scale_colour_tvc_c()
}
