% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_tvc_d.R
\name{scale_colour_tvc_d}
\alias{scale_colour_tvc_d}
\title{Discrete Colour Scale}
\usage{
scale_colour_tvc_d(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{either gradient or "main" or ?}

\item{direction}{1 for onward or -1 for downward}

\item{...}{other params ?}
}
\value{
\code{scale_colour_tvc_d} returns the discrete tvc colour palette for points and lines.
}
\description{
Description
}
\examples{
scale_colour_tvc_d()
}
