% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_tvc_d.R
\name{scale_fill_tvc_d}
\alias{scale_fill_tvc_d}
\title{Discrete Fill Scale}
\usage{
scale_fill_tvc_d(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{either gradient or "main" or ?}

\item{direction}{1 for onward or -1 for downward}

\item{...}{other params ?}
}
\value{
\code{scale_fill_tvc_d} returns the discrete tvc fill palette for areas.
}
\description{
Description
}
\examples{
scale_fill_tvc_d()
}
