% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_coverage.R
\name{triad_coverage}
\alias{triad_coverage}
\title{Triad Coverage}
\usage{
triad_coverage(triads)
}
\arguments{
\item{triads}{a tibble triad}
}
\value{
\code{triad_coverage} returns a tibble
}
\description{
Triad Coverage
}
\examples{
triads <- read_sheet_triad()
triad_coverage(triads)
}
