% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvc-package.R
\docType{package}
\name{tvc-package}
\alias{tvc}
\alias{tvc-package}
\title{tvc: Vanessa Carlton Triad}
\description{
R functions to analyse Vanessa Carlton Triads through music history.
}
\author{
\strong{Maintainer}: Arthur Petit \email{arthur.petit5108@gmail.com}

}
\keyword{internal}
