% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_missing_region.R
\name{write_missing_region}
\alias{write_missing_region}
\title{Write Missing Region}
\usage{
write_missing_region(songs)
}
\arguments{
\item{songs}{a tibble of songs x artists as returned by \code{\link{get_spotify_api_playlist}}}
}
\value{
\code{write_missing_region} write new artists with missing region in the triad google sheet.
}
\description{
Write Missing Region
}
\examples{
# write_missing_region(songs)
}
