% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_missing_triade.R
\name{write_missing_triade}
\alias{write_missing_triade}
\title{Update Triad Google Sheet}
\usage{
write_missing_triade(songs, triads, ss = NULL, triadeType = "1 7")
}
\arguments{
\item{songs}{a tibble of songs from Spotify as returned by \code{\link{get_spotify_api_playlist}}}

\item{triads}{a tibble of triads from Google Sheet "Triade"}

\item{ss}{a google sheet identifier}

\item{triadeType}{a triad identifier, "1 7" for Vanessa Carlton triad}
}
\value{
\code{write_missing_triade} writes the missing songs in a google sheet "Missing Triade" so that the user can copy and paste them to the main "Triade" sheet
}
\description{
Update Triad Google Sheet
}
\examples{
triads <- read_sheet_triad()
playlist_id <- get_playlist_id(triad = "1 7")
songs <- tvc:::get_spotify_api_playlist(playlist_id = playlist_id, scope = "public") \%>\% 
  dplyr::mutate(triad = "1 7")
write_missing_triade(songs = songs, triads = triads)
}
