% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spotify_api_triad_playlists.R
\name{get_spotify_api_triad_playlists}
\alias{get_spotify_api_triad_playlists}
\title{Get All Triads}
\usage{
get_spotify_api_triad_playlists()
}
\value{
\code{get_spotify_api_triad_playlists} returns a tibble of songs for different triads
}
\description{
A Wrapper around \code{\link{get_spotify_api_playlist}} to get all triad's songs at once in an optimized way
}
\examples{
# get_spotify_api_triad_playlists()
}
