% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_coverage.R
\name{triad_coverage}
\alias{triad_coverage}
\title{Triad Coverage}
\usage{
triad_coverage(triads, weighted = TRUE)
}
\arguments{
\item{triads}{a tibble triad}

\item{weighted}{a boolean specifying if songs should be weighted using tvc's bias correction method}
}
\value{
\code{triad_coverage} returns a tibble
}
\description{
Triad Coverage
}
\examples{
triads <- read_sheet_triad()
dat <- triad_coverage(triads = triads)
}
